backend/requirements.txt
Django>=3.2,<4.2.10
psycopg[binary]
daphne==4.1.0
djangorestframework==3.14.0
requests